Simulations
These are codes which read the Excel file that contains the simulated payments between banks and carries out the simulation.  The simulated data are generated in the other file attached.  The simulation of one sample is fast, however, the iteration is very large.  I would like to ask if I can submit it to CUDA.  As each iteration looks heavy in space complexity, I don't know if it is doable.  Any comments on improving the code are welcomed!